Move request logging into `DevServer`
